Add verbosity flag to the root command
