+++ Sources/swift-bundler/main.swift
@@ -1,22 +1,4 @@
 import Foundation
-import ArgumentParser
-import Darwin
-
-/// The root command of Swift Bundler.
-struct SwiftBundler: ParsableCommand {
-  static let configuration = CommandConfiguration(
-    commandName: "swift-bundler",
-    abstract: "A tool for creating macOS apps from Swift packages.",
-    version: "v2.0.0-rc.1",
-    shouldDisplay: true,
-    subcommands: [
-      BundleCommand.self,
-      RunCommand.self,
-      CreateCommand.self,
-      TemplatesCommand.self,
-      GenerateXcodeSupportCommand.self
-    ])
-}
 
 // Kill all running processes on exit
 #if os(macOS)
@@ -25,7 +7,7 @@ for signal in Signal.allCases {
     for process in processes {
       process.terminate()
     }
-    exit(1)
+    Foundation.exit(1)
   }
 }
 #endif
